Chore: Move CategoryPopoverContent to a separate folder
