+++ src/components/CategoryPopover/CategoryPopoverContent/index.js
@@ -1,0 +1,1 @@
+export { default } from './CategoryPopoverContent';
